+++ tests/testapp/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class TestAppConfig(AppConfig):
+    default = True
+    name = "tests.testapp"
+    label = "testapp"
+    verbose_name = "Guardrails test app"